Minor changes to PrimaryDrive, renamed AutoOne to MainAuto, and deleted an unused opmode called AutoDrive. Made the Autonomous mode run significantly faster, in order to prepare for scoring multiple samples in Autonomous.
